integer overflow on Windows in datetime transformer v3

+++ tsururu/transformers/datetime.py
@@ -77,7 +77,7 @@ class TimeToNumGenerator(FeaturesGenerator):
             _, time_delta = index_slicer.timedelta(time_col, delta=self.delta)
             if self.from_target_date:
                 horizon = data["target_idx"][0, -1] - data["features_idx"][0, -1]
-                time_col = time_col + np.int64(horizon) * np.int64(time_delta)
+                time_col = time_col + np.int64(horizon) * time_delta
 
             new_arr = pd.to_datetime(time_col.to_numpy().reshape(-1), origin="unix")
             data_transformed = (
@@ -181,7 +181,7 @@ class DateSeasonsGenerator(FeaturesGenerator):
             _, time_delta = index_slicer.timedelta(time_col, delta=self.delta)
             if self.from_target_date:
                 horizon = data["idx_y"][0, -1] - data["idx_X"][0, -1]
-                time_col = time_col + np.int64(horizon) * np.int64(time_delta)
+                time_col = time_col + np.int64(horizon) * time_delta
 
             time_col = pd.to_datetime(time_col.to_numpy(), origin="unix")
